feat: refactor tests package
