+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/expression/ResolveFeelExpressionTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.processing.feelexpressionresolution;
+package io.camunda.zeebe.engine.processing.expression;
 
 import static org.assertj.core.api.Assertions.assertThat;
